refactor: change package
utils moved to package specific due include util tests
